Move head to app.vue
